add products model and handler

+++ migrations/sqls/20220204134004-products-table-up.sql
@@ -2,6 +2,6 @@
 CREATE TABLE products (
     id integer PRIMARY KEY generated always as identity,
     name VARCHAR(100) NOT NULL,
-    price DECIMAL(13,2) NOT NULL,
+    price integer NOT NULL,
     category VARCHAR(50) NOT NULL
 );
